Split `ProxyMixin` out into its own file
This:
- Makes the separation between `ProxyBase` and the legacy `ProxyMixin`
  clearer.
- Makes it possible to import cuml without sklearn installed. Once
  sklearn is a required dependency (and the `ProxyMixin` is removed),
  `is_proxy` can be moved back into `estimator_proxy.py` where it makes
  more sense to be located.
