+++ python/cuml/cuml/accel/_wrappers/sklearn/cluster.py
@@ -15,7 +15,7 @@
 #
 
 import cuml.cluster
-from cuml.accel.estimator_proxy import ProxyMixin
+from cuml.accel.estimator_proxy_mixin import ProxyMixin
 
 __all__ = ("KMeans", "DBSCAN")
