+++ python/cuml/cuml/accel/_wrappers/sklearn/manifold.py
@@ -15,7 +15,7 @@
 #
 
 import cuml.manifold
-from cuml.accel.estimator_proxy import ProxyMixin
+from cuml.accel.estimator_proxy_mixin import ProxyMixin
 
 __all__ = ("TSNE",)
