+++ python/cuml/cuml/accel/_wrappers/sklearn/neighbors.py
@@ -15,7 +15,7 @@
 #
 
 import cuml.neighbors
-from cuml.accel.estimator_proxy import ProxyMixin
+from cuml.accel.estimator_proxy_mixin import ProxyMixin
 
 __all__ = ("NearestNeighbors", "KNeighborsClassifier", "KNeighborsRegressor")
